--- tests/Rector/Contrib/Symfony/Console/CommandToConstructorInjectionRector/config.yml
@@ -1,9 +1,0 @@
-rectors:
-    Rector\Rector\Contrib\Symfony\Console\CommandToConstructorInjectionRector: ~
-
-services:
-    Rector\Tests\Rector\Contrib\Symfony\Console\CommandToConstructorInjectionRector\Source\DummyProvider: ~
-
-    # require for interface autowiring
-    Rector\Contract\Bridge\ServiceTypeForNameProviderInterface:
-        alias: Rector\Tests\Rector\Contrib\Symfony\Console\CommandToConstructorInjectionRector\Source\DummyProvider
