+++ tests/Rector/Contrib/Symfony/FrameworkBundle/ContainerGetToConstructorInjectionRector/config.yml
@@ -1,0 +1,9 @@
+rectors:
+    Rector\Rector\Contrib\Symfony\FrameworkBundle\ContainerGetToConstructorInjectionRector: ~
+
+services:
+    Rector\Tests\Rector\Contrib\Symfony\FrameworkBundle\ContainerGetToConstructorInjectionRector\Source\DummyProvider: ~
+
+    # require for interface autowiring
+    Rector\Contract\Bridge\ServiceTypeForNameProviderInterface:
+        alias: Rector\Tests\Rector\Contrib\Symfony\FrameworkBundle\ContainerGetToConstructorInjectionRector\Source\DummyProvider
